Updated basket calculations and added transport API call
Updated the basket to use a full version of jQuery instead of a slim version to be able to handle AJAX call. Included the VignobleWEB.Core.Models namespace in Basket.cshtml to use the application model. Updated the dropdown menu on the basket page so it now sends the selected transport's ID to the backend, which is needed for the new transport price calculation. Then, added an AJAX API call that retrieves details of the selected transport option from the backend, including its price. This is used to correctly calculate the total price. For this call, an authentication token is also used which is added in Basket.cshtml.cs. Updated styling and added error handling for mentioned AJAX call. Also, made minimal changes in _Layout.cshtml to ensure jQuery version and integrity. In the Basket.cshtml.cs backend file, the token from the TokenManager is obtained. Plus, changed the way transport options are added to the dropdown menu: now using transport's ID instead of price as a value. Updated the Transport model class to map correctly with the backend data model.